pyproject.toml
[project]
name = "autoreadme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "together>=1.5.35",
]
